requirements.txt
beautifulsoup4==4.10.0
openpyxl==3.0.9
pandas==1.3.5
pymssql==2.2.4
python-dotenv==0.19.2
requests==2.31.0
rich==13.7.1
SQLAlchemy==1.4.32
tablepyxl==0.6.1
xlrd==2.0.1
humanfriendly==10.0
Babel==2.14.0
pendulum>=2.1.2
playwright>=1.35.0
PyYAML>=6.0.1
python-dateutil>=2.8.1